Eric's original HashMap method, redone. You'll notice my method is a lot shorter than his, because I did read through his code and there was plenty of redundancy. Also he had to worry about many formatting errors that we no longer have to worry for.
Checking for new instructor data is easy since we just check for an Empty line. 
That is how Eric was doing it before; however, his code was disorganized and could have been tremendously cut down in terms of lines and length. 

I do want you guys to look it over and see if there could be any improvements or changes made.
The way I tested the code was on a much smaller scale with a few classes I had made seperately. I will upload those.